# swagger.yaml
openapi: 3.0.0
info:
  title: MindLoo API
  version: 2.0.0
  description: |
    Complete RESTful API for managing patients, therapists, appointments, and therapy-related operations.
    
    ## Features
    - Patient and Therapist Management
    - Appointment Scheduling
    - Therapy Session Tracking
    - Medical Records Management
    - Secure Messaging System
    - Review and Rating System
    - Treatment Planning and Goal Tracking
    
  contact:
    name: Stefan Ramač
    url: https://www.linkedin.com/in/stefanramac
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  schemas:
...
    Appointment:
      type: object
      required:
        - therapistId
        - patientId
        - date
        - timeSlot
      properties:
        appointmentId:
          type: string
          readOnly: true
        therapistId:
          type: string
          example: "T001"
        patientId:
          type: string
          example: "P001"
        date:
          type: string
          format: date
          example: "2025-10-25"
        timeSlot:
          type: object
          required:
            - start
            - end
          properties:
            start:
              type: string
              example: "10:00"
            end:
              type: string
              example: "11:00"
        status:
          type: string
          enum: [scheduled, completed, cancelled, rescheduled]
          default: scheduled
        subject:
          type: string
          example: "Initial consultation"
        notes:
          type: string
    
    # Availability Schema
...
paths:
...
  /api/appointments:
    post:
      tags:
        - Appointments
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '201':
          description: Appointment created successfully
        '400':
          description: Time slot not available
    
    get:
      tags:
        - Appointments
      summary: Get appointments with filters
      parameters:
        - name: therapistId
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, completed, cancelled, rescheduled]
      responses:
        '200':
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
  /api/appointments/{id}:
    get:
      tags:
        - Appointments
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Appointment data
        '404':
          description: Appointment not found
    
    patch:
      tags:
        - Appointments
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                subject:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Appointment updated successfully
    
    delete:
      tags:
        - Appointments
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Appointment deleted successfully
  
  # Availability Endpoints
...